bot/app/requirements.txt
matrix-nio[e2e]
requests
vaderSentiment
